dry_run argument added to JamfPackageCleaner
An optional dry_run argument has been added. If set to true, nothing will be deleted from Jamf Pro.

Use together with -v to see a list of packages that would be kept/deleted.

+++ JamfUploaderProcessors/JamfPackageCleaner.py
@@ -68,6 +68,13 @@ class JamfPackageCleaner(JamfUploaderBase):
             "This is used as a failsafe.",
             "default": "20",
         },
+            "dry_run": {
+            "required": False,
+            "description": "If set to True, nothing is deleted from Jamf Pro. "
+            "Use together with '-v' for detailed information. "
+            "This is used for testing",
+            "default": False,
+        },
     }
 
     output_variables = {
@@ -177,6 +184,7 @@ class JamfPackageCleaner(JamfUploaderBase):
         self.maximum_allowed_packages_to_delete = int(
             self.env.get("maximum_allowed_packages_to_delete")
         )
+        self.dry_run = self.env.get("dry_run")
 
         # Create a list of smb shares in tuples
         self.smb_shares = []
@@ -256,7 +264,7 @@ class JamfPackageCleaner(JamfUploaderBase):
                 "Override by changing the 'minimum_name_length' argument. Aborting."
             )
             return
-
+    
         # Get all packages from Jamf Pro as JSON object
         self.output(f"Getting all packages from {self.jamf_url}")
         token, send_creds, _ = self.handle_classic_auth(
@@ -296,10 +304,20 @@ class JamfPackageCleaner(JamfUploaderBase):
                 "Override by setting the 'maximum_allowed_packages_to_delete' argument. Aborting."
             )
             return
-
-        # Print the packages to keep and delete
+        
+        #  Print the packages to keep and delete
         for package in packages_to_keep:
-            self.output(f"Keeping {package['name']}", verbose_level=2)
+            self.output(f"✅ {package['name']}", verbose_level=1)
+
+        for package in packages_to_delete:
+            self.output(f"❌ {package['name']} (will be deleted)", verbose_level=1)
+    
+        # If running dry, print intentions and abort.
+        if self.dry_run:
+            self.output(f"INFO: Argument 'dry_run' is set to True. Nothing will be deleted. "
+                        f"Use '-v' to see detailed information. "
+                        f"Aborting.")
+            return
 
         for package in packages_to_delete:
             self.delete_package(
